fix build issues for x64 and x86
